+++ apps/dashboard/src/components/settings/ApiKeys/Alerts.tsx
@@ -2,7 +2,9 @@ import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
 import { TrackedUnderlineLink } from "@/components/ui/tracked-link";
 import { CircleAlertIcon } from "lucide-react";
 
-export const SmartWalletsBillingAlert = () => {
+export const SmartWalletsBillingAlert = (props: {
+  teamSlug: string;
+}) => {
   return (
     <Alert variant="warning">
       <CircleAlertIcon className="size-5" />
@@ -10,7 +12,7 @@ export const SmartWalletsBillingAlert = () => {
       <AlertDescription>
         To enable AA on mainnet chains,{" "}
         <TrackedUnderlineLink
-          href="/team/~/~/settings/billing"
+          href={`/team/${props.teamSlug}/~/settings/billing`}
           category="api_keys"
           label="smart_wallets_missing_billing"
         >