+++ packages/thirdweb/src/react/core/hooks/pay/useBuyWithCryptoStatus.ts
@@ -52,6 +52,7 @@ import {
  *  return <button onClick={handleBuyWithCrypto}>Swap</button>
  * }
  * ```
+ * @deprecated
  * @buyCrypto
  */
 export function useBuyWithCryptoStatus(params?: BuyWithCryptoTransaction) {